FIX: Prevent multi-base REF/ALT from matching SNPs
ref/alt match when ref/alt is "in" possible matches. However when iupac
is false, possible matches is never converted into a set, so the string
representation is instead searched. That means, for example, the value
"A" will be considered to match "TA".

Set ref/alt to {ref/alt} when iupac is false.