FIX: Don't count variants multiple times
If a variant appears in multiple parts of a query, it is counted once
for each time it appears.

Only count the variant once. As each search that finds a variant in a
given dataset should find it the same number of times, we can just use
the last value found.