REFACTOR: Increase recursion factor
The delay between a function being called for the first time and
processing being complete is larger than anticipated, which means the
recursion factor must be increased. The optimum recursion factor was
calculated as f(ln(f)-1) = d/r, where f is the optimum recursion factor,
d is the length of time the last completed function took to complete
less the actual runtime of the associated lambda function, and r is the
total runtime of the function minus d divided by the number of functions
run (i.e. the rate at which functions were called, in seconds per
function).

Over a range of query types (32-variant profile and spike protein range
query) d/r was found to be around the range 194-211, leading to an
optimum recursion factor of 62-66, so 64 was chosen.