FIX: Call summariseDataset from one summariseSlice
Once a vcf is completely summarised, the last summariseSlice calls
summariseDataset again. This means that if a dataset is composed of many
files, each one will cause summariseDataset to be called again.

Mark all files that need to be updated before summariseDataset can be
called again. This means only the final updated file will call
summariseDataset. As a side effect, this means that other datasets that
contain these files will no longer be automatically updated, as the
toUpdateFiles field will not have been set in dynamoDB. Calling
submitDataset with other datasets will cause them to be updated.