FIX: Remove ID from query and backend
summariseSampleMetadata is running out of memory, and a large consumer
of memory and time is the handling of accession IDs.

Since nobody seems to be using this feature, just remove the ID field
from the queries and sample metadata. This will buy us time until
perhaps 13 million samples, at which point we'll have to look at better
optimisations for summariseSampleMetadata.