FIX: Increase timeout for summariseDataset
summariseDataset is timing out between checking the etag of every file
and calling summariseVcf for every file.

Increase the timeout to two minutes, which should be enough to handle up
to 20 million samples (2048 files). If there are still problems, calling
submit with a subset of the files first will allow summariseVcf to be
called in batches (or just increase the timeout again).