FEAT: Add login variable to base terraform
The website module has a terraform variable to set the login config
setting, but this variable is not exposed to the main terraform script.

Create a variable that exposes the possibility to change the login
variable in the same place that the other variables are set, and change
the value to false as we are now serving public data.

Note that even with this variable set to false, the website still
appears to redirect to the login page, so further work is required to
make it publicly accessible.